Async: allow appending existing Job objects to the Job chain
Now it's possible to do something like

Job<int, int> job = createSomeJob();
auto main = Async::start<int>(....).then(job);

Previously the 'job' would have to be wrapped in a ThenTask-like lambda (which
is what we still do internally), but with this new syntax it's possible to append
another job chain to existing chain easilly. This syntax is available for all
task types.