Immediately close the socket after a shutdown notification.
Otherwise it can happen that the synchronizer receives the shutdown
command from a ResourceAccess instance, and while the synchronizer
waits for quit to be called, a new ResourceAccess instance already
connects to the synchronizer. That new connection will then of course
be unexpectedly killed once quit is called in the synchronizer.