Cache ResourceAccess instances.
We want one connection per application per resource instance.
With this change the connection is maintained for the lifetime of the
client process (it should probably time out instead), but we at least
avoid creating a connection per operation/query, which results in
a significant performance boost (~10% for reading, 90% for writing in
dummyresourcebenchmark).