Improved imap mail sync algorithm.
* when requesting individual mails we sync the full content
* when requesting individual folders we get 2 weeks of full content +
  headers for everything else.
* when requesting a sync for all folders we only get 2 weeks of full
  content.

Getting the headers for 50k messages takes about 180s on my system with
kolabnow (network being the bottleneck), so that's managable. Getting
the full content would take in the range of hours.
This way we have something to show, and a way to request more data,
without making the system overly complex yet.
Certainly not the final solution, but a good start.