Support dependencies between sync requests.
If one sync task depends on the previous sync task we want to flush in
between, so we can query for the results of the previous sync request
locally.

If we detect such a dependency we temporarily halt all processing of
synchronization requests until the flush completes, so we can continue
processing.