Don't store the lockfile inside the main resource folder.
We currently use the folder as the main db folder, meaning we remove
the folder when removing the db. This results in the lockfile vanishing
with the db, which then confuses QLockFile (resulting in a lot of
warnings). We may want to start moving everything into a
resource-instance folder, but then we have to do it properly across the
board.