Use read-write locks for finer grained control to sDbi and sEnvironments
There are only a few cases where have to access the list of dbis or
environments, so we can normally get away with just read-locking.

This seems to fix a segfault that was possibly caused be an environment
being reused that has already been freed in another thread. The
read-only lock when initially retrieving the environment seems to fix
that.