Fixed crashes due to concurrently running queries.
A single QueryRunner should never have multiple workers running at the
same time. We did not properly enforce this in case of incremental
updates coming in.

The only way I managed to reproduce the crash:
* Open a large folder with lots of unread mail in kube
* Select a mail in the maillist and hold the down button
* This will:
    * Repeatedly call fetch more
    * Trigger lot's of mark as read modifications that result in
    notifications.
    * Eventually it crashes somewhere in EntityStore, likely because
    of concurrent access of the filter structure which is shared through
    the state.

We now ensure in the single threaded portion of the code that we only
ever run one worker at a time. If we did receive an update during,
we remember that change and fetch more once we're done.

To be able to call fetch again that portion was also factored out into a
separate function.