Create dbis using a child transaction that we immediately commit.
We now protect dbi initialization with a mutex and immediately update
the sDbis hash. This assumes that the created dbis are indeed
immediately valid for other transactions (before the parent transaction
commits).

Opening dbis on environment init is furhter separated out, so we don't
end up in the regular openDatabase codepath at all.