Optimize item fetching in DAV protocols
Summary:
When updating a collection, fetch every items in that collection in one go, instead of querying the server once for every items.

In the future, it would be nice to construct a URL list composed of items that has a different ETag than the one in the store, instead of just every items.

Reviewers: cmollekopf

Tags: #sink

Differential Revision: https://phabricator.kde.org/D15058