Added the .vcf and .ics suffices for webdav.
iRony has a bug in that it assumes the suffixes are always used (it
shouldn't). Because we can't work around that bug in any other way
(If we list items we wouldn't know if now should or should not remove
the suffix), and since it seems to be standard practice anyways,
we'll just include the suffix when creating new items.