Only abort changereplay on transient errors
Changereplay will stop and block any further changes,
so the only chase where we should stop replaying is if we
are rather certain that the error will go away (e.g. when the network is
back). In all other cases we have to move on and deal with the
consequences.