Only commit after 100 messages instead of 10
One implication of the previous commit is that as soon as we commit to
the messagequeue and yield to the eventloop, the pipeline takes it's
turn and processes the just comitted messages.
This has the unfortunate effect that we no longer process batches of
100 as defined by the batchsize in commandprocessor, but instead just
batches of 10, which greatly reduces performance.

The reason is of course that producer and consumer are not properly
decoupled, but the quick workaround is to always commit after 100
messages, which happens to match the batchsize we are looking for.