Removed mailman support
We don't even have a testmessage and I don't think it's very useful for
our usecase.

It seems to be implementing a custom mailman format to turn a plaintext
message back into a list of messages, but mailman in principle supports
delivering digests with proper MIME included messages (just like a
forwarded message), so that seems like the better option.