Only deal with enabled specialpurpose folders.
We used to sometimes start appending sent messages to a disabled folder
(if two sent folders are present), which has the surprising effect that
your messages vanish.

The downside of this approach might be that we try to create a folder
that is already existing but disabled, which we may have to handle
separately if it becomes a problem.