Avoid removing local data due to a temporary error.
I ran into this due to a "Error while opening transaction: Resource temporarily unavailable" error.
I'm not sure what the source of that error was, but the result was that
the upgrade code removed the local database, so I had to resync all my
data.

This should prevent that.

ChangeReplay later on crashed, so we now handle that more explicitly.