Avoid setting unnecessarily lowe lower bound revisions
* We would set it to 0 even if we don't query anything.
* We are actually setting it to 0 even if it was higher before,
so it's just 0 most of the time preventing cleanup (unless all clients
also do livequeries).

The proper way to do it would be to:
* Find the revision we're about to read.
* Protect that revision by sending the command and ensuring it has been
processed (and potentially try again if it's gone).
* Only then actually read (which then guarantees that all future
incremental updates will have all revisions available)

But that's going to be fairly complicated for something that is only
relevant for read access (no dataloss), and is very unlikely to happen
(famous last words).