Avoid relying on default encoding
We now try harder to respect the charset of the mail, and avoid just
using kmime's default encoding (ISO-8859-1).

To that end we have to make sure we pass through the encoding from
parent parts to child parts.
Also, when creating a temporary node the parents encoding still
applies, so we have to set it.

With this all tests pass again also with the latest kmime (which started
to explicitly set the default charset to us-ascii, breaking some tests).