Actually send revision replayed commands
After sending a mail the mailtransport resource
quickly initiated the move to sentfolder (and thus removal from the resource).
This removal was immediately replayed, and then cleand up because we
failed to send revision replayed commands to the resource.
This in turned caused an assert in kube's outbox query, because
we tried to replay a revision that was no longer existing.