Also verify signatures in the mailpreprocessor.
For non-detached signatures, the signed content becomes only available when we
process the signature.
In principle we could of course provide a fast-path for detached
signatures which simply calls parseInternal in the messagepart on the content to make it
available.

Let's see if this simpler solution is performant enough or not.