DAV "conflict resolution"
We simply overwrite our local version.

So far we had the problem that we protect local modifications in the
pipeline, which resulted in the local version being preserved, but
server and client being prepetually out of sync (Server couldn't
overwrite local state, and client couldn't update server because of ctag).

The alternative approaches would be:
* Merge state somehow
* Create a new item
* Overwrite the server state instead