Set the correct flags to scan over keys with an empty key
The only thing we needed to do is to position the cursor at the initial
item, instead of trying to set it to the (empty) key.

The operations we used before happened to work in that specific case,
but lead to unexpected results in other cases.

This is a fixup for a49a6a2baad8b89e46f2def18303ec349ba94d2a