Treat empty and absent strings the same
In old buffers we don't have the parentUid available, which resulted in
a null string aka QVariant{}. In the new buffers the parentUid is set to
an empty string and there is no way to remove it.
We now instead always treat an empty string as no string at all, so it
will result in an invalid QVariant.

The downside of this approach is that we can't query for "", just for
QVariant{} (with the current comparator anyways).